pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arquimedes"
version = "0.1.0"
description = "Collaborative LLM knowledge base for architecture practice and research"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pymupdf>=1.24",
    "pdfplumber>=0.11",
    "anthropic>=0.40",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "jinja2>=3.1",
    "watchdog>=5.0",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
arq = "arquimedes.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
